# Build `serde` and `serde_derive` in parallel (22% faster builds)