--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,7 +19,8 @@ rustdoc-args = ["--cfg", "docsrs"]
 [dev-dependencies]
 hex = "0.4.2"
 serde_json = "1.0.41"
-serde = { version = "1.0", features = ["derive"] }
+serde = "1.0"
+serde_derive = "1.0"
 
 [dependencies.ed25519-compact]
 version = "2.0.2"