Before:

    $ hyperfine --prepare "cargo clean" 'cargo build --tests'
    Benchmark 1: cargo build --tests
      Time (mean ± σ):      6.546 s ±  0.153 s    [User: 20.745 s, System: 3.745 s]
      Range (min … max):    6.241 s …  6.853 s    10 runs

After:

    $ hyperfine --prepare "cargo clean" 'cargo build --tests'
    Benchmark 1: cargo build --tests
      Time (mean ± σ):      5.100 s ±  0.117 s    [User: 21.400 s, System: 3.917 s]
      Range (min … max):    4.956 s …  5.276 s    10 runs

The version mismatch problem that prohibited doing this earlier was fixed almost [2 years
ago](https://github.com/serde-rs/serde/issues/2584#issuecomment-2938307179).

(Not doing this prevents downstream projects from performing the same build time optimization since cargo features are
[additive](https://doc.rust-lang.org/cargo/reference/resolver.html#features) across the dependency graph.)